--- application/api/util/validator/alarm_validator_test.go
@@ -1,1 +1,0 @@
-package validator
