--- application/api/util/validator/validator_test.go
@@ -1,1 +1,0 @@
-package validator
