+++ application/infrastructure/database/chara.go
