--- application/infrastructure/database/database_user_ios_platform_info.go
@@ -1,8 +1,0 @@
-package database
-
-type UserIOSPlatformInfo struct {
-	PushToken                string `dynamodbav:"pushToken"`
-	PushTokenSNSEndpoint     string `dynamodbav:"pushTokenSNSEndpoint"`
-	VoIPPushToken            string `dynamodbav:"voIPPushToken"`
-	VoIPPushTokenSNSEndpoint string `dynamodbav:"voIPPushTokenSNSEndpoint"`
-}
